feat: separate footer
