--- .github/pull_request_template.md
@@ -1,32 +1,0 @@
-<!--
-    PR name should have the following convention:       type(scope): message | FNC-XYZ
-    
-    For multiple commits for isolated change,
-    on merge, squash commits with message:              type(scope): message (#pr-number) | FNC-XYZ
-
-    If PR is still in progress, you can add WIP in the beginning of PR name.
--->
-
-## Description
-
-- [ ] My Pull Request Bugfix or Hotfix
-- [ ] My Pull Request Feature
-
-<!--
-    Please include a short summary of the change and which issue is fixed. Please also include relevant motivation and context.
--->
-
-
-## How Has This Been Tested?
-
-<!--
-    Please describe the tests that you ran to verify your changes.
- -->
-
--
-
-## Checklist
-
-- [ ] I have performed a self-review of my code
-- [ ] My changes generate no new console warnings or errors
-- [ ] My PR had no conflict and able to merge
