--- src/Features/moleculs/DashboardFooter/DashboardFooter.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-
-import styles from './dashboardFooter.module.scss';
-
-const DashboardFooter: React.FC = () => {
-  return <div className={styles.dashboardFooter}>DashboardFooter</div>;
-};
-export default DashboardFooter;
