--- src/Features/moleculs/DashboardFooter/dashboardFooter.module.scss
@@ -1,13 +1,0 @@
-@import "/src/styles/vars";
-@import "/src/styles/mixins";
-.dashboardFooter {
-    height: 50px;
-    margin-left: 380px;
-    width: calc(100% - 460px);
-    z-index: 1;
-    border-top: 4px solid rgba(182, 193, 255, 0.12);
-    @media all and (max-width: $normalDesktop) {
-        margin-left: 320px;
-        width: calc(100% - 376px);
-    } 
-}
