--- src/Features/moleculs/DashboardFooter/index.ts
@@ -1,3 +1,0 @@
-import DashboardFooter from './DashboardFooter';
-
-export default DashboardFooter;
